+++ webserver/static/js/feverscreen.ts
@@ -2,6 +2,7 @@ import {BlobReader} from "./utils.js";
 import {fahrenheitToCelsius, moduleTemperatureAnomaly, sensorAnomaly, ROIFeature} from "./processing.js";
 import {DeviceApi} from "./api.js";
 import {CalibrationInfo, FrameInfo, Modes, NetworkInterface, TemperatureSource} from "./feverscreen-types.js";
+import {circleDetect, circleDetectRadius, edgeDetect} from "./circledetect.js";
 import {buildSAT, scanHaar, ConvertCascadeXML, HaarCascade} from "./haarcascade.js";
 
 let GROI : ROIFeature[] = [];
@@ -771,124 +772,6 @@ window.onload = async function() {
     return dest;
   }
 
-  function addPixel(dest: Float32Array, x: number, y: number, amount: number, width: number, height: number) {
-    x=~~x
-    y=~~y
-    if(x<0 || y<0) { return;}
-    if(x>=width || y>=height) { return;}
-    const index = y * width + x;
-    dest[index] += amount;
-  }
-
-  function addCircle(dest: Float32Array, cx: number, cy: number, radius: number, amount: number, width: number, height: number) {
-    addPixel(dest, cx+radius, cy, amount, width, height);
-    addPixel(dest, cx-radius, cy, amount, width, height);
-    addPixel(dest, cx, cy+radius, amount, width, height);
-    addPixel(dest, cx, cy-radius, amount, width, height);
-    let d=3-(2*radius);
-    let ix=1;
-    let iy=radius;
-    while(ix<iy) {
-        //Bresenham
-        if(d<0) {
-            d += 4*ix+6;
-        }else{
-            iy=iy-1;
-            d+=4*(ix-iy)+10;
-        }
-        addPixel(dest, cx+ix, cy+iy, amount, width, height);
-        addPixel(dest, cx-ix, cy+iy, amount, width, height);
-        addPixel(dest, cx+ix, cy-iy, amount, width, height);
-        addPixel(dest, cx-ix, cy-iy, amount, width, height);
-        addPixel(dest, cx+iy, cy+ix, amount, width, height);
-        addPixel(dest, cx-iy, cy+ix, amount, width, height);
-        addPixel(dest, cx+iy, cy-ix, amount, width, height);
-        addPixel(dest, cx-iy, cy-ix, amount, width, height);
-        ix+=1;
-    }
-  }
-
-  function edgeDetect(source: Float32Array) {
-    const width = frameWidth;
-    const height = frameHeight;
-    const dest = new Float32Array(width * height);
-    for (let y = 2; y < height-2; y++) {
-      for (let x = 2; x < width-2; x++) {
-        let index = y * width + x;
-        let value = source[index]*4-source[index-1]-source[index+1] - source[index+1*width] - source[index-1*width];
-        dest[index] = Math.max(value-40,0);
-      }
-    }
-
-    return dest;
-  }
-
-  function circleDetectRadius(
-        source: Float32Array, dest: Float32Array, radius: number, width: number, height: number,
-        wx0 : number, wy0 : number, wx1 : number, wy1 : number
-        ): number[] {
-    radius=Math.max(radius,0.00001)
-    for(let i=0; i<width * height; i++) {
-        dest[i] = 0;
-    }
-
-    wx0 = Math.max(wx0, 2);
-    wy0 = Math.max(wy0, 2);
-    wx1 = Math.min(wx1, width - 2);
-    wy1 = Math.min(wy1, height - 2);
-
-    for (let y = wy0; y < wy1; y++) {
-      for (let x = wx0; x < wx1; x++) {
-        let index = y * width + x;
-        let value = source[index]
-        if(value < 1) { continue; }
-        addCircle(dest, x, y, radius, 1, width, height);
-      }
-    }
-    let result = 0;
-    let rx=0;
-    let ry=0;
-    for (let y = wy0; y < wy1; y++) {
-      for (let x = wx0; x < wx1; x++) {
-        let index = y * width + x;
-        if(result < dest[index]) {
-          result = dest[index];
-          rx = x;
-          ry = y;
-        }
-      }
-    }
-    return [result/(2+radius), rx, ry];
-  }
-
-  function circleDetect(source: Float32Array): [Float32Array, number, number, number] {
-    const width = frameWidth;
-    const height = frameHeight;
-    const dest = new Float32Array(width * height);
-    let radius=3.0;
-    let bestRadius = -1;
-    let bestValue = 2;
-    let bestX=0;
-    let bestY=0;
-
-    while(radius<20){
-        let value=0;
-        let cx=0;
-        let cy=0;
-        [value, cx, cy] = circleDetectRadius(source, dest, radius, width, height, 2, 2, width-2, height - 2);
-        if(bestValue < value) {
-          bestValue = value;
-          bestRadius = radius;
-          bestX=cx;
-          bestY=cy;
-        }
-        radius = ~~(radius * 1.03 + 1);
-    }
-
-    //circleDetectRadius(source, dest, bestRadius, width, height);
-    return [dest, bestRadius, bestX, bestY];
-  }
-
   function mip_scale_down(source: Float32Array, width: number, height: number): Float32Array {
     const ww = Math.floor(width / 2);
     const hh = Math.floor(height / 2);
@@ -1109,8 +992,8 @@ window.onload = async function() {
   }
 
   function detectThermalReference(roi : ROIFeature[], saltPepperData: Float32Array, smoothedData: Float32Array, width : number, height : number) {
- //   const edgeData = edgeDetect(saltPepperData);
-    const edgeData = edgeDetect(smoothedData);
+ //   const edgeData = edgeDetect(saltPepperData, frameWidth, frameHeight);
+    const edgeData = edgeDetect(smoothedData, frameWidth, frameHeight);
 
     if(GROI.length>0) {
         let prevTherm = GROI[GROI.length-1];
@@ -1125,7 +1008,7 @@ window.onload = async function() {
     let bestRadius;
     let bestX;
     let bestY;
-    [circle_image, bestRadius, bestX, bestY] = circleDetect(edgeData);
+    [circle_image, bestRadius, bestX, bestY] = circleDetect(edgeData, frameWidth, frameHeight);
 
     if (bestRadius<=0) {
       return roi;
